requirements.txt
streamlit
feedparser
transformers
newspaper3k
torch
